refactor(common): deduplicate chunk operation implementations
