+++ src/common/src/array/stream_chunk.rs
@@ -24,6 +24,7 @@ use rand::prelude::SmallRng;
 use rand::{Rng, SeedableRng};
 use risingwave_pb::data::{PbOp, PbStreamChunk};
 
+use super::stream_chunk_builder::StreamChunkBuilder;
 use super::{ArrayImpl, ArrayRef, ArrayResult, DataChunkTestExt, RowRef};
 use crate::array::DataChunk;
 use crate::buffer::{Bitmap, BitmapBuilder};
@@ -182,33 +183,20 @@ impl StreamChunk {
     /// For consecutive `UpdateDelete` and `UpdateInsert`, they will be kept in one chunk.
     /// As a result, some chunks may have `size + 1` rows.
     pub fn split(&self, size: usize) -> Vec<Self> {
-        let data_types = self.data_types();
-        let mut rows = Vec::with_capacity(size + 1);
-        let mut results = vec![];
+        let mut builder = StreamChunkBuilder::new(size, self.data_types());
+        let mut outputs = Vec::new();
 
-        let mut iter = self.rows();
-        while let Some(row) = iter.next() {
-            rows.push(row);
-            if rows.len() == size {
-                // If the last row is UpdateDelete, also include the UpdateInsert.
-                if rows.last().unwrap().0 == Op::UpdateDelete {
-                    let next_row = iter
-                        .next()
-                        .expect("UpdateDelete should have UpdateInsert after");
-                    assert_eq!(next_row.0, Op::UpdateInsert);
-                    rows.push(next_row);
-                }
-                let chunk = Self::from_rows(&rows, &data_types);
-                results.push(chunk);
-                rows.clear();
+        // TODO: directly append the chunk.
+        for (op, row) in self.rows() {
+            if let Some(chunk) = builder.append_row(op, row) {
+                outputs.push(chunk);
             }
         }
-
-        if !rows.is_empty() {
-            let chunk = StreamChunk::from_rows(&rows, &data_types);
-            results.push(chunk);
+        if let Some(output) = builder.take() {
+            outputs.push(output);
         }
-        results
+
+        outputs
     }
 
     pub fn into_parts(self) -> (DataChunk, Arc<[Op]>) {
@@ -534,38 +522,10 @@ impl StreamChunkMut {
         }
     }
 }
+
 /// Test utilities for [`StreamChunk`].
-pub trait StreamChunkTestExt: Sized {
-    fn from_pretty(s: &str) -> Self;
-
-    /// Validate the `StreamChunk` layout.
-    fn valid(&self) -> bool;
-
-    /// Concatenate multiple `StreamChunk` into one.
-    fn concat(chunks: Vec<Self>) -> Self;
-
-    /// Sort rows.
-    fn sort_rows(self) -> Self;
-
-    /// Generate stream chunks
-    fn gen_stream_chunks(
-        num_of_chunks: usize,
-        chunk_size: usize,
-        data_types: &[DataType],
-        varchar_properties: &VarcharProperty,
-    ) -> Vec<Self>;
-
-    fn gen_stream_chunks_inner(
-        num_of_chunks: usize,
-        chunk_size: usize,
-        data_types: &[DataType],
-        varchar_properties: &VarcharProperty,
-        visibility_percent: f64, // % of rows that are visible
-        inserts_percent: f64,
-    ) -> Vec<Self>;
-}
-
-impl StreamChunkTestExt for StreamChunk {
+#[easy_ext::ext(StreamChunkTestExt)]
+impl StreamChunk {
     /// Parse a chunk from string.
     ///
     /// See also [`DataChunkTestExt::from_pretty`].
@@ -606,7 +566,7 @@ impl StreamChunkTestExt for StreamChunk {
     /// //   x[]: array of x
     /// // {i,f}: struct
     /// ```
-    fn from_pretty(s: &str) -> Self {
+    pub fn from_pretty(s: &str) -> Self {
         let mut chunk_str = String::new();
         let mut ops = vec![];
 
@@ -647,34 +607,35 @@ impl StreamChunkTestExt for StreamChunk {
         }
     }
 
-    fn valid(&self) -> bool {
+    /// Validate the `StreamChunk` layout.
+    pub fn valid(&self) -> bool {
         let len = self.ops.len();
         let data = &self.data;
         data.visibility().len() == len && data.columns().iter().all(|col| col.len() == len)
     }
 
-    fn concat(chunks: Vec<StreamChunk>) -> StreamChunk {
-        assert!(!chunks.is_empty());
-        let mut ops = vec![];
-        let mut data_chunks = vec![];
-        let mut capacity = 0;
+    /// Concatenate multiple `StreamChunk` into one.
+    ///
+    /// Panics if `chunks` is empty.
+    pub fn concat(chunks: Vec<StreamChunk>) -> StreamChunk {
+        let data_types = chunks[0].data_types();
+        let size = chunks.iter().map(|c| c.cardinality()).sum::<usize>();
+
+        let mut builder = StreamChunkBuilder::new(size + 1, data_types);
+
         for chunk in chunks {
-            capacity += chunk.capacity();
-            ops.extend(chunk.ops.iter());
-            data_chunks.push(chunk.data);
+            // TODO: directly append chunks.
+            for (op, row) in chunk.rows() {
+                let none = builder.append_row(op, row);
+                debug_assert!(none.is_none());
+            }
         }
-        let data = DataChunk::rechunk(&data_chunks, capacity)
-            .unwrap()
-            .into_iter()
-            .next()
-            .unwrap();
-        StreamChunk {
-            ops: ops.into(),
-            data,
-        }
+
+        builder.take().expect("empty chunk")
     }
 
-    fn sort_rows(self) -> Self {
+    /// Sort rows.
+    pub fn sort_rows(self) -> Self {
         if self.capacity() == 0 {
             return self;
         }
@@ -693,7 +654,7 @@ impl StreamChunkTestExt for StreamChunk {
     /// Generate `num_of_chunks` data chunks with type `data_types`,
     /// where each data chunk has cardinality of `chunk_size`.
     /// TODO(kwannoel): Generate different types of op, different vis.
-    fn gen_stream_chunks(
+    pub fn gen_stream_chunks(
         num_of_chunks: usize,
         chunk_size: usize,
         data_types: &[DataType],
@@ -709,7 +670,7 @@ impl StreamChunkTestExt for StreamChunk {
         )
     }
 
-    fn gen_stream_chunks_inner(
+    pub fn gen_stream_chunks_inner(
         num_of_chunks: usize,
         chunk_size: usize,
         data_types: &[DataType],
